Reorganizing /app to remove /src directory.
