+++ app/app.js
@@ -5,8 +5,8 @@ const queryString = require('querystring');
 const rateLimit = require('express-rate-limit');
 const url = require('url');
 
-require('dotenv').config({ path: path.resolve(__dirname, '../../.env') });
-const { BASE_URL, EIA_KEY, TEST_KEY } = require('./constants/constants');
+require('dotenv').config({ path: path.resolve(__dirname, '../.env') });
+const { BASE_URL, EIA_KEY, TEST_KEY } = require('./constants');
 
 const app = express();
 const baseController = require('./controllers/baseController');
@@ -51,7 +51,7 @@ app.get('/', async (req, res) => res.send(await baseController(TEST_KEY)));
  */
 
 // EIA category smoke test
-app.get('/api/categories/test', async (req, res) => {
+app.get('v1/api/categories/test', async (req, res) => {
   /* TODO: Eliminate this test route when real routes are complete */
   try {
     const results = await categoriesTestController(BASE_URL, EIA_KEY, '714755');
@@ -62,7 +62,7 @@ app.get('/api/categories/test', async (req, res) => {
   }
 });
 
-app.get('/api/categories/data', async (req, res) => {
+app.get('v1/api/categories/data', async (req, res) => {
   /* FIXME: Add express-validator to prevent XSS */
   try {
     const category_id = await req.query.category_id;
@@ -85,7 +85,7 @@ app.get('/api/categories/data', async (req, res) => {
  */
 
 // EIA series smoke test
-app.get('/api/series/test', async (req, res) => {
+app.get('v1/api/series/test', async (req, res) => {
   /* TODO: Eliminate this test route when real routes are complete */
   try {
     const results = await seriesTestController(BASE_URL, EIA_KEY);
@@ -97,7 +97,7 @@ app.get('/api/series/test', async (req, res) => {
 });
 
 // EIA series with query parameters
-app.get('/api/series/data', async (req, res) => {
+app.get('v1/api/series/data', async (req, res) => {
   /* FIXME: Add express-validator to prevent XSS */
   try {
     const series_id = await req.query.series_id;
