+++ app/server.js
@@ -1,4 +1,4 @@
-const app = require('./src/app');
+const app = require('./app');
 const port = 3000;
 
 app.listen(port, () => {
